Added Readme
**************************************************************************
Ada Web Server Files
====================

Because it's so troublesome to compile the system for AWS (Ada Wrb
Server) on Windows, requiring that you pollute your environment with a
unix-like subsystem (Cygwin, MSys, etc), I'm posting the results of the
compile/install here.

Built with:
-----------
* Gnat 2012 (GPL)
* AdaXML
* OpenSSL

_Notes:_
--------

* The file _AWS-Basic.zip_ contains the contents of my base-folder which
had
been modified.
* The file _AWS-with-Directories.zip_ contains everything Basic does,
plus the directories that had been modified (and their contents).

**************************************************************************